# docs/api/scribe.openapi.yaml
openapi: 3.1.0
info:
  title: Scribe Connect API
  version: v1
  description: Workspace-scoped APIs for handwritten-document ingest, transcription, IIIF annotation editing, and publication.
paths:
  /scribe.v1.AnnotationService/EnrichAnnotation:
    post:
      tags:
        - scribe.v1.AnnotationService
      summary: Improve transcription for one line or a full canvas.
      description: Improve transcription for one line or a full canvas.
      operationId: scribe.v1.AnnotationService.EnrichAnnotation
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.EnrichAnnotationRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.EnrichAnnotationResponse'
  /scribe.v1.AnnotationService/ExportAnnotationPage:
    post:
      tags:
        - scribe.v1.AnnotationService
      summary: Export one exact committed canonical revision to a rebuildable OCR view.
      description: Export one exact committed canonical revision to a rebuildable OCR view.
      operationId: scribe.v1.AnnotationService.ExportAnnotationPage
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.ExportAnnotationPageRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.ExportAnnotationPageResponse'
  /scribe.v1.AnnotationService/GetAnnotation:
    post:
      tags:
        - scribe.v1.AnnotationService
      summary: GetAnnotation
      operationId: scribe.v1.AnnotationService.GetAnnotation
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.GetAnnotationRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.GetAnnotationResponse'
  /scribe.v1.AnnotationService/GetAnnotationPage:
    post:
      tags:
        - scribe.v1.AnnotationService
      summary: GetAnnotationPage
      operationId: scribe.v1.AnnotationService.GetAnnotationPage
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.GetAnnotationPageRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.GetAnnotationPageResponse'
  /scribe.v1.AnnotationService/JoinLines:
    post:
      tags:
        - scribe.v1.AnnotationService
      summary: JoinLines
      operationId: scribe.v1.AnnotationService.JoinLines
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.JoinLinesRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.JoinLinesResponse'
  /scribe.v1.AnnotationService/JoinWordsIntoLine:
    post:
      tags:
        - scribe.v1.AnnotationService
      summary: JoinWordsIntoLine
      operationId: scribe.v1.AnnotationService.JoinWordsIntoLine
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.JoinWordsIntoLineRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.JoinWordsIntoLineResponse'
...
  /scribe.v1.AnnotationService/SaveAnnotationPage:
    post:
      tags:
        - scribe.v1.AnnotationService
      summary: SaveAnnotationPage
      operationId: scribe.v1.AnnotationService.SaveAnnotationPage
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.SaveAnnotationPageRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.SaveAnnotationPageResponse'
  /scribe.v1.AnnotationService/SearchAnnotations:
    post:
      tags:
        - scribe.v1.AnnotationService
      summary: SearchAnnotations
      operationId: scribe.v1.AnnotationService.SearchAnnotations
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.SearchAnnotationsRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.SearchAnnotationsResponse'
  /scribe.v1.AnnotationService/SplitLineIntoTwoLines:
    post:
      tags:
        - scribe.v1.AnnotationService
      summary: SplitLineIntoTwoLines
      operationId: scribe.v1.AnnotationService.SplitLineIntoTwoLines
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.SplitLineIntoTwoLinesRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.SplitLineIntoTwoLinesResponse'
  /scribe.v1.AnnotationService/SplitLineIntoWords:
    post:
      tags:
        - scribe.v1.AnnotationService
      summary: Structural split/join operations.
      description: Structural split/join operations.
      operationId: scribe.v1.AnnotationService.SplitLineIntoWords
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.SplitLineIntoWordsRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.SplitLineIntoWordsResponse'
  /scribe.v1.AnnotationService/SplitPageIntoWords:
    post:
      tags:
        - scribe.v1.AnnotationService
      summary: SplitPageIntoWords
      operationId: scribe.v1.AnnotationService.SplitPageIntoWords
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.SplitPageIntoWordsRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.SplitPageIntoWordsResponse'
...
  /scribe.v1.ImageProcessingService/ProcessHOCR:
    post:
      tags:
        - scribe.v1.ImageProcessingService
      summary: ProcessHOCR
      operationId: scribe.v1.ImageProcessingService.ProcessHOCR
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.ProcessHOCRRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.ProcessHOCRResponse'
...
  /scribe.v1.ItemService/PrepareItemExport:
    post:
      tags:
        - scribe.v1.ItemService
      summary: PrepareItemExport
      operationId: scribe.v1.ItemService.PrepareItemExport
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.PrepareItemExportRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.PrepareItemExportResponse'
...
  /scribe.v1.ItemService/UploadItemImage:
    post:
      tags:
        - scribe.v1.ItemService
      summary: UploadItemImage
      operationId: scribe.v1.ItemService.UploadItemImage
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/scribe.v1.UploadItemImageRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scribe.v1.UploadItemImageResponse'
...
components:
  schemas:
    connect-protocol-version:
      type: number
      title: Connect-Protocol-Version
      enum:
        - 1
      description: Define the version of the Connect protocol
      const: 1
    connect-timeout-header:
      type: number
      title: Connect-Timeout-Ms
      description: Define the timeout, in ms
    connect.error:
      type: object
      properties:
        code:
          type: string
          examples:
            - not_found
          enum:
            - canceled
            - unknown
            - invalid_argument
            - deadline_exceeded
            - not_found
            - already_exists
            - permission_denied
            - resource_exhausted
            - failed_precondition
            - aborted
            - out_of_range
            - unimplemented
            - internal
            - unavailable
            - data_loss
            - unauthenticated
          description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
        message:
          type: string
          description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
        details:
          type: array
          items:
            $ref: '#/components/schemas/connect.error_details.Any'
          description: A list of messages that carry the error details. There is no limit on the number of messages.
      title: Connect Error
      additionalProperties: true
      description: 'Error type returned by Connect: https://connectrpc.com/docs/go/errors/#http-representation'
...
    scribe.v1.EnrichAnnotationRequest:
      type: object
      properties:
        itemImageId:
          exclusiveMinimum: 0
          type:
            - integer
            - string
          title: item_image_id
          format: int64
          description: |-
            Exact tenant-scoped image whose region will be transcribed. Canvas IDs are
             targets/provenance and are not authorization or lookup keys.
        scope:
          type: string
          title: scope
          pattern: .*\S.*
          description: '"line" = single Annotation JSON; "page" = full AnnotationPage JSON'
        annotationJson:
          type: string
          title: annotation_json
          pattern: .*\S.*
        contextId:
          type:
            - integer
            - string
          title: context_id
          format: int64
          description: 'Optional: override context used for transcription'
      title: EnrichAnnotationRequest
      required:
        - itemImageId
      additionalProperties: false
      description: |
        EnrichAnnotationRequest sends a IIIF annotation (one line or an entire
         AnnotationPage) to Scribe's transcription service for improvement.
         The caller provides the raw JSON and specifies the scope.
        enrich_annotation.scope // item_image_id and annotation_json are required, and scope must be line or page
    scribe.v1.EnrichAnnotationResponse:
      type: object
      properties:
        annotationJson:
          type: string
          title: annotation_json
      title: EnrichAnnotationResponse
      additionalProperties: false
    scribe.v1.ExportAnnotationPageRequest:
      type: object
      properties:
        itemImageId:
          exclusiveMinimum: 0
          type:
            - integer
            - string
          title: item_image_id
          format: int64
        expectedRevision:
          exclusiveMinimum: 0
          type:
            - integer
            - string
          title: expected_revision
          format: int64
          description: |-
            The exact committed revision to export. A stale value fails explicitly so
             callers can never mistake a newer background result for their saved draft.
        format:
          not:
            enum:
              - 0
          title: format
          $ref: '#/components/schemas/scribe.v1.AnnotationExportFormat'
      title: ExportAnnotationPageRequest
      required:
        - itemImageId
        - expectedRevision
        - format
      additionalProperties: false
      description: |
        export_annotation_page.required // item_image_id, expected_revision, and format are required
    scribe.v1.ExportAnnotationPageResponse:
      type: object
      properties:
        itemImageId:
          type:
            - integer
            - string
          title: item_image_id
          format: int64
        revision:
          type:
            - integer
            - string
          title: revision
          format: int64
        mediaType:
          type: string
          title: media_type
        content:
          type: string
          title: content
          format: byte
        filename:
          type: string
          title: filename
      title: ExportAnnotationPageResponse
      additionalProperties: false
...
    scribe.v1.GetAnnotationRequest:
      type: object
      properties:
        itemImageId:
          exclusiveMinimum: 0
          type:
            - integer
            - string
          title: item_image_id
          format: int64
        id:
          type: string
          title: id
          pattern: .*\S.*
          description: Complete canonical Scribe annotation ID from the current page.
      title: GetAnnotationRequest
      required:
        - itemImageId
        - id
      additionalProperties: false
    scribe.v1.GetAnnotationResponse:
      type: object
      properties:
        annotationJson:
          type: string
          title: annotation_json
      title: GetAnnotationResponse
      additionalProperties: false
...
    scribe.v1.JoinLinesResponse:
      type: object
      properties:
        annotationPageJson:
          type: string
          title: annotation_page_json
          description: Complete transformed AnnotationPage JSON.
      title: JoinLinesResponse
      additionalProperties: false
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: scribe_session
      description: Interactive browser session cookie. The cookie name is deployment-configurable; scribe_session is the default.
    ScribeAPIKey:
      type: apiKey
      in: header
      name: X-Scribe-API-Key
      description: Workspace-scoped Scribe API key.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Scribe API key or registered external JWT
security:
  - SessionCookie: []
  - ScribeAPIKey: []
  - BearerAuth: []
...